requirements.txt
# Core Flask (unchanged)
flask==2.3.3
werkzeug==2.3.7
python-dotenv==0.19.0

# Updated LangChain (add Gemini support)
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5

# Google Gemini (replacing openai)
google-generativeai==0.3.2

# Keep your existing vector store
chromadb==0.4.0

# Document processing (unchanged)
pypdf==3.9.0
unstructured==0.10.30

# Azure Storage (replacing boto3/S3)
azure-storage-blob==12.19.0
azure-identity==1.15.0